+++ pkg/exams/R/fileURI.R
@@ -41,6 +41,14 @@
   "xlsx","application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
   "zip", "application/zip",
 
+  "mp3", "audio/mpeg",
+  "ogg", "audio/ogg",
+  "flac","audio/flac",
+  "wav", "audio/wav",
+  "avi", "video/x-msvideo",
+  "mp4", "video/mp4",
+  "mkv", "video/x-matroska",
+
   "gh",  "application/octet-stream",
   "3dm", "model/vnd.3dm"
 ), ncol = 2L, byrow = TRUE, dimnames = list(NULL, c("ext", "mime")))